ADDED logic for auto selection of section icon based on the components used.
Based on the provided panels data, determine what a good automatic icon
for this section would be. The default is returned if there's no components
or we're unsure what to do.
If there's a quiz anywhere, we use that icon.
If there's only one component, we use the icon for that component.
If there's multiple components, and one has more instances than the others,
we use the icon for that component.
Otherwise we use the icon for the first component in the list.